requirements-test.txt
# Dependências para testes
pytest>=7.0.0
pytest-dotenv>=0.5.2
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
python-dotenv>=1.0.0

# Dependências já existentes do projeto
requests>=2.31.0
supabase>=1.0.0
openai>=1.0.0

# Novas dependências adicionadas
pytest
requests
python-dotenv
supabase 